pyproject.toml
[project]
name = "hse-applied-python-fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "asyncpg>=0.29.0", 
    "redis>=4.5.0",
    "python-dotenv>=1.2.2",
]
